공부 리스트 뷰모델 만들다가 중간 커밋
