+++ Gisellina/Gisellina/Views/MissionListView/StudyMissionListView.swift
@@ -9,12 +9,13 @@ import SwiftUI
 
 struct StudyMissionListView: View {
     @EnvironmentObject var router: Router
+    let missionList: [MissionList]
     
     var body: some View {
         ScrollView {
             VStack(spacing: 14) {
-                ForEach(0..<6) { _ in
-                    StudyMissionListCell()
+                ForEach(missionList) { mission in
+                    StudyMissionListCell(missionDetail: mission)
                 }
             }
             .padding(.top, 20)
@@ -23,34 +24,6 @@ struct StudyMissionListView: View {
     }
 }
 
-struct StudyMissionListCell: View {
-    var body: some View {
-        HStack(spacing: 10){
-            VStack(alignment: .leading, spacing: 4) {
-                Text("2025.06.01")
-                    .font(.system(size: 14, weight: .regular))
-                    .foregroundStyle(.gray)
-                Text("Q. 증여자가 수증기의 범죄행위를 원인으로 하여 증여계약을 수증자가 그...")
-                    .font(.system(size: 16, weight: .bold))
-            }
-            Spacer()
-//                        Image(systemName: "checkmark.circle.fill")
-//                            .resizable()
-//                            .frame(width: 40, height: 40)
-            Image(systemName: "chevron.right")
-                .font(.system(size: 24, weight: .bold))
-                .foregroundStyle(.blue .opacity(0.5))
-        }
-        .padding(24)
-        .background(Color.blue.opacity(0.1))
-        .cornerRadius(20)
-        .overlay(
-            RoundedRectangle(cornerRadius: 20)
-                .stroke(Color.blue.opacity(0.5), lineWidth: 1)
-            )
-    }
-}
-
-#Preview {
-    StudyMissionListView()
-}
+//#Preview {
+//    StudyMissionListView()
+//}
